improve memory helper functions
