+++ tests/common/file_utils_test.py
@@ -1,4 +1,3 @@
-from datetime import timedelta
 from collections import Counter
 import os
 import pathlib
@@ -20,8 +19,6 @@ from allennlp.common.file_utils import (
     open_compressed,
     CacheFile,
     _Meta,
-    format_size,
-    format_timedelta,
     _find_entries,
     inspect_cache,
     remove_cache_entries,
@@ -452,35 +449,3 @@ class TestCacheFile(AllenNlpTestCase):
                 raise IOError("I made this up")
         assert not os.path.exists(handle.name)
         assert not os.path.exists(cache_filename)
-
-
-@pytest.mark.parametrize(
-    "size, result",
-    [
-        (12, "12B"),
-        (int(1.2 * 1024), "1.2K"),
-        (12 * 1024, "12K"),
-        (120 * 1024, "120K"),
-        (int(1.2 * 1024 * 1024), "1.2M"),
-        (12 * 1024 * 1024, "12M"),
-        (120 * 1024 * 1024, "120M"),
-        (int(1.2 * 1024 * 1024 * 1024), "1.2G"),
-        (12 * 1024 * 1024 * 1024, "12G"),
-    ],
-)
-def test_format_size(size: int, result: str):
-    assert format_size(size) == result
-
-
-@pytest.mark.parametrize(
-    "td, result",
-    [
-        (timedelta(days=2, hours=3), "2 days"),
-        (timedelta(days=1, hours=3), "1 day"),
-        (timedelta(hours=3, minutes=12), "3 hours"),
-        (timedelta(hours=1, minutes=12), "1 hour, 12 mins"),
-        (timedelta(minutes=12), "12 mins"),
-    ],
-)
-def test_format_timedelta(td: timedelta, result: str):
-    assert format_timedelta(td) == result
