+++ tests/training/trainer_test.py
@@ -107,7 +107,7 @@ class TestTrainer(TrainerTestBase):
         assert "best_epoch" in metrics
         assert isinstance(metrics["best_epoch"], int)
         assert "peak_worker_0_memory_MB" in metrics
-        assert isinstance(metrics["peak_worker_0_memory_MB"], float)
+        assert isinstance(metrics["peak_worker_0_memory_MB"], int)
         assert metrics["peak_worker_0_memory_MB"] > 0
 
     def test_trainer_can_run_exponential_moving_average(self):
@@ -130,7 +130,7 @@ class TestTrainer(TrainerTestBase):
         )
         metrics = trainer.train()
         assert "peak_worker_0_memory_MB" in metrics
-        assert isinstance(metrics["peak_worker_0_memory_MB"], float)
+        assert isinstance(metrics["peak_worker_0_memory_MB"], int)
         assert metrics["peak_worker_0_memory_MB"] > 0
         assert "peak_gpu_0_memory_MB" in metrics
         assert isinstance(metrics["peak_gpu_0_memory_MB"], int)
